Alemeno/credit_approval/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
celery>=5.3.4
redis>=5.0.1
pandas>=2.1.3
openpyxl>=3.1.2
gunicorn>=21.2.0
python-dateutil>=2.8.2
